fix: Split the API file
