+++ src/store/reducers.ts
@@ -1,17 +1,23 @@
 import { combineReducers } from "@reduxjs/toolkit";
 import userSlice from "@/store/slices/userSlice";
-import { appApi } from "@/store/api";
 import upworkFeedSlice from "@/store/slices/upworkFeedSlice";
 import upworkFeedDetailSlice from "@/store/slices/upworkFeedDetailSlice";
 import allChatsSlice from "@/store/slices/allChatsSlice";
 import messagesSlice from "@/store/slices/messagesSlice";
+import { authApi } from "@/store/authApi";
+import { upworkApi } from "@/store/upworkApi";
+import { chatsApi } from "@/store/chatsApi";
+import { messagesApi } from "@/store/messagesApi";
 
 const rootReducer = combineReducers({
-  [appApi.reducerPath]: appApi.reducer,
   user: userSlice,
   upworkFeed: upworkFeedSlice,
   upworkFeedDetail: upworkFeedDetailSlice,
   allChats: allChatsSlice,
   chat: messagesSlice,
+  [authApi.reducerPath]: authApi.reducer,
+  [upworkApi.reducerPath]: upworkApi.reducer,
+  [chatsApi.reducerPath]: chatsApi.reducer,
+  [messagesApi.reducerPath]: messagesApi.reducer,
 });
 export default rootReducer;
