+++ src/views/UpworkFeed/UpworkFeed.tsx
@@ -1,8 +1,8 @@
 import { ReactElement, useEffect } from "react";
-import { useUpworkFeedsQuery } from "@/store/api";
 import RSSRefreshBar from "@/views/UpworkFeed/RSSRefreshBar/RSSRefreshBar";
 import UpworkFeedTable from "@/views/UpworkFeed/UpworkFeedTable/UpworkFeedTable";
 import useSearchParameters from "@/hooks/useSearchParameters";
+import { useUpworkFeedsQuery } from "@/store/upworkApi";
 
 function UpworkFeed(): ReactElement {
   const { pageSize, pageNumber, searchParameters, sortDirection } =
